Move page-size spec tests to the existing core test file
Finish the content_size_bytes -> unpadded_page_size_bytes rename in the
source (the prior commit renamed it only in the test) and relocate the
tests into tests/v1/core/test_kv_cache_utils.py, reusing new_kv_cache_spec.
